+++ src/components/Settings/index.ts
@@ -1,0 +1,1 @@
+export { default as SettingsMain } from "./SettingsMain";
